Organized directory and added comments
