+++ src/slimeimgcreator.py
